+++ pizza/templates/edicao.html
@@ -26,7 +26,7 @@
                     <input class="form-control" id="url_foto" name="url_foto" value="{{pizza['url_foto']}}">
                 </div>
                 <div class="form-group col-lg-2 col-md-4 col-sm-12 col-12">
-                   <div class="img-pizza" style="background-image: url({{pizza['url_foto']}});"></div>
+                   <div class="img-pizza" style="background-image: url('{{pizza['url_foto']}}');"></div>
                 </div>
                 <div class="form-group col-lg-12 col-md-12 col-sm-12 col-12">
                     <label for="descricao"> Descrição: </label>
